fix(pnl): 신규 항목 손익분류 기본값을 구획에서 파생 + 금액 컬럼 numeric
수입 구획에 만든 항목이 변동비로 잡혀 부호가 뒤집히던 문제를 고친다.
구획→분류 파생은 신규 항목 초기값에만 적용한다(둘은 독립 축).
금액 컬럼을 스키마 관례대로 numeric으로 맞춘다(미적용 마이그레이션).
계획서의 "DDL 전 클라우드는 잔액 0" 서술이 틀려 정정한다 — 실제로는
시드 부트스트랩이 돌아 정상처럼 보이면서 편집이 조용히 유실된다.

+++ supabase/migrations/0018_plan_items.sql
@@ -7,11 +7,11 @@ create table if not exists plan_items (
   id text primary key,
   "group" text not null check ("group" in ('income','spending','saving')),
   name text not null,
-  amount integer not null,
+  amount numeric not null,
   pnl_class text not null check (pnl_class in ('revenue','fixed','saving','variable')),
   conditional boolean not null default false,
   end_year_month text,
-  target_total integer,
+  target_total numeric,
   note text,
   sort_order integer not null default 0,
   created_at timestamptz not null default now()
